Redoes file structure This commit redoes the file structure to sort the files by type as recommended by React (https://legacy.reactjs.org/docs/faq-structure.html)
fisk-events-board/
├── public/
│   ├── index.html
│   ├── manifest.json
│   └── assets/
│       ├── images/
│       └── icons/
├── src/
│   ├── assets/
│   │   ├── images/
│   │   └── icons/
│   ├── components/
│   ├── contexts/
│   ├── hooks/
│   ├── pages/
│   ├── services/
│   ├── utils/
│   ├── App.js
│   ├── index.js
│   ├── index.css
│   └── routes.js
├── .gitignore
├── package.json
├── README.md
└── package-lock.json

Yet to fully organize files correctly. Will do in next commit
